backend/requirements.txt

fastapi
uvicorn
openai
pyjwt
